add SchedulerUserService to handle different users; no details yet

+++ src/main/java/org/anstreth/schedulebot/model/User.java
@@ -1,10 +1,9 @@
 package org.anstreth.schedulebot.model;
 
-import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
 public class User {
     final long id;
-    final long groupId;
+    final int groupId;
 }
